Merged PR 1389: Create basic platforming logic & encapsulate sprites and text into classes
The weirdest part of this is the CRC code, which essentially lets us refer to sprites by their path without incurring the cost of string lookups all the time. Aside from that, I've tried to encapsulate most of the SDL2 code into the renderer and sprite classes, so that the game logic part of the code doesn't have to deal with that.

Movement is kinda in there - although it only moves 1 pixel when the key is pressed. We need some way to store the current input device state, so that repeated actions (movement) trigger every frame.

We also need a way of getting frame timings so that refresh rate doesn't affect movement speed.